fractional delay support

+++ include/atae/delay/delay_line.h
@@ -8,8 +8,8 @@ public:
     float processSample(float input);
     void reset();
 private:
-    int delay_samples = 0;
-    int max_delay_samples = 0;
+    float delay_samples = 0.0f;
+    float max_delay_samples = 0.0f;
     int sample_rate = 0;
     int write_index = 0;
     std::vector<float> delay_buffer;
